+++ zappend/slice/__init__.py
@@ -3,9 +3,10 @@
 # https://opensource.org/licenses/MIT.
 
 from .abc import SliceSource
+from .factory import SliceCallable
 from .factory import SliceFactory
 from .factory import SliceObj
 from .factory import open_slice_source
 from .factory import to_slice_factories
 from .factory import to_slice_factory
-from .factory import to_slice_source_type
+from .factory import to_slice_callable