+++ zappend/slice/memory.py
@@ -4,7 +4,6 @@
 
 import xarray as xr
 
-from ..context import Context
 from ..log import logger
 from .abc import SliceSource
 
@@ -13,13 +12,11 @@ class MemorySliceSource(SliceSource):
     """A slice source that uses the in-memory dataset passed in.
 
     Args:
-        ctx: Processing context
         slice_ds: The slice dataset
         slice_index: An index for slice identification (logging only)
     """
 
-    def __init__(self, ctx: Context, slice_ds: xr.Dataset, slice_index: int):
-        super().__init__(ctx)
+    def __init__(self, slice_ds: xr.Dataset, slice_index: int):
         self._slice_ds = slice_ds
         self._slice_index = slice_index
 
@@ -30,4 +27,3 @@ class MemorySliceSource(SliceSource):
     def dispose(self):
         self._slice_ds = None
         logger.info(f"Slice dataset #{self._slice_index} processed")
-        super().dispose()